Move ApiId to Constants
