--- lib/msal-browser/src/telemetry/ApiId.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export enum ApiId {
-    acquireTokenRedirect = 861,
-    acquireTokenPopup = 862,
-    ssoSilent = 863,
-    acquireTokenSilent_authCode = 864,
-    handleRedirectPromise = 865,
-    acquireTokenSilent_silentFlow = 61
-};
